+++ engine/src/main/java/nl/arkenbout/geoffrey/angel/engine/core/graphics/util/Vector4u.java
@@ -1,10 +1,10 @@
 package nl.arkenbout.geoffrey.angel.engine.core.graphics.util;
 
-import org.joml.Vector3f;
-import org.joml.Vector4f;
+import org.joml.Vector3d;
+import org.joml.Vector4d;
 
 public class Vector4u {
-    public static Vector3f xyz(Vector4f v) {
-        return new Vector3f(v.x(), v.y(), v.z());
+    public static Vector3d xyz(Vector4d v) {
+        return new Vector3d(v.x(), v.y(), v.z());
     }
 }
